Remove port comparison from same_origin()
`same_origin` returns False if the referrer port and current host are not equal, resulting in a Bad Request - Referrer checking failed - origin does not match.

This is problematic when Nginx is proxy passing requests to a Gunicorn application running Flask. Nginx is running on one port, and Gunicorn on another.